debug folding

+++ modeling/modeling.py
@@ -102,12 +102,12 @@ class RebaseT5(pl.LightningModule):
         # alternatively, you can do this in __init__: [for parameter in self.esm1b_model.paramters(): parmater.requires_grad=False]
         # pass that ESM-1b hidden states into self.model(..., encoder_outputs=...)
         
-        output = self.model(encoder_outputs=[batch['embedding']], attention_mask=mask, labels=batch['bind'])
+        output = self.model(encoder_outputs=[batch['embedding']], attention_mask=mask, labels=batch['bind'].long())
 
         
         
         self.log('train_loss', float(output.loss), on_step=True, on_epoch=True, prog_bar=True, logger=True)
-        self.log('train_acc',float(accuracy(output['logits'].argmax(-1), batch['bind'], (batch['bind'] != -100).int())), on_step=True, on_epoch=True, prog_bar=False, logger=True)
+        self.log('train_acc',float(accuracy(output['logits'].argmax(-1), batch['bind'].long(), (batch['bind'] != -100).int())), on_step=True, on_epoch=True, prog_bar=False, logger=True)
         # self.log('train_perplex',float(self.perplexity(output['logits'], batch['bind'])), on_step=True, on_epoch=True, prog_bar=False, logger=True)
         
         return {
@@ -124,9 +124,9 @@ class RebaseT5(pl.LightningModule):
         # 1 for tokens that are not masked; 0 for tokens that are masked
         mask = (batch['seq'] != self.dictionary.pad()).int()
         with torch.no_grad():
-            output = self.model(encoder_outputs=[batch['embedding']], attention_mask=mask, labels=batch['bind'])
+            output = self.model(encoder_outputs=[batch['embedding']], attention_mask=mask, labels=batch['bind'].long())
         self.log('val_loss', float(output.loss), on_step=True, on_epoch=True, prog_bar=False, logger=True)
-        self.log('val_acc',float(accuracy(output['logits'].argmax(-1), batch['bind'], (batch['bind'] != self.dictionary.pad()).int())), on_step=True, on_epoch=True, prog_bar=False, logger=True)
+        self.log('val_acc',float(accuracy(output['logits'].argmax(-1), batch['bind'].long(), (batch['bind'] != self.dictionary.pad()).int())), on_step=True, on_epoch=True, prog_bar=False, logger=True)
         return {
             'loss': output.loss,
             'batch_size': batch['seq'].size(0)
